% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano.R
\name{Volcano}
\alias{Volcano}
\title{This is a volcano plot function}
\usage{
Volcano(
  data = data,
  x = "log2FoldChange",
  y = "Padj",
  fill,
  color = "Threshold",
  color_palette = c("#8DA0CB", "grey", "#FC8D62"),
  label = "gene",
  label_data = subset_data,
  alpha = 0.5,
  plotLab = T,
  xlab = "log2 Fold Change",
  ylab = "log10 Padj",
  title = "Volcano Plot of genes",
  xline = c(-1, 1)
)
}
\arguments{
\item{data}{the dataframe you would like to display}

\item{x}{the colname of data which you would like to use as x value.}

\item{y}{the colname of data which you would lite to use as y value.}

\item{fill}{the colname of data you would like to draw color(fill color)}

\item{color}{the colname of data you would like to draw color(edge color)}

\item{color_palette}{all colors your would like to use in your figure.}

\item{label}{the lable name you would like to put}

\item{label_data}{A dataframe which you would like to take label}

\item{alpha}{the transparency of points}

\item{plotLab}{the logical value, if you would like to display label text.}

\item{xlab}{the name of x-asix, with the default string of "log2FoldChange".}

\item{ylab}{the name of y-asix, with the default string of "-log10 Padj".}

\item{title}{the name of this figure. default name is "Volcano Plot of genes"}

\item{xline}{the xintercept, default is c(-1,1).}
}
\value{
a volcano plot of ggplot format
}
\description{
This is a volcano plot function
}
